ADD: Images and Patch README.md

+++ Core/Source/Core/Chip8.cpp
@@ -11,7 +11,7 @@
 
 // character sprite = 5 bytes;
 
-uint8_t fontset[FONTSET_SIZE] = {
+u8 fontset[FONTSET_SIZE] = {
 	0xF0, 0x90, 0x90, 0x90, 0xF0, // 0
 	0x20, 0x60, 0x20, 0x20, 0x70, // 1
 	0xF0, 0x10, 0xF0, 0x80, 0xF0, // 2
@@ -34,7 +34,7 @@ Chip8::Chip8():randGen(std::chrono::system_clock::now().time_since_epoch().count
 
     pc = START_ADDRESS;
 
-    for(unsigned int i = 0; i < FONTSET_SIZE; ++i){
+    for(size_t i = 0; i < FONTSET_SIZE; ++i){
         memory[FONTSET_START_ADDRESS + i] = fontset[i];
     }
